激光采用5V 0.7% 控制; 激光识别上加入了动态阈值, 提高识别精度

+++ opencv_laser/LaserCtrlor.h
@@ -11,6 +11,6 @@ public:
 	void setduty(int key);
 private:
 	CSerialPort *comport;
-	int duty;
+	double duty;
 };
